Replace payment_methods cols with json('details')
