+++ src/Models/PaymentMethod.php
@@ -6,29 +6,22 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use rkujawa\LaravelPaymentGateway\Database\Factories\PaymentMethodFactory;
-use rkujawa\LaravelPaymentGateway\PaymentGateway;
+use rkujawa\LaravelPaymentGateway\Models\Traits\PaymentMethodRequests;
+
+=
 
 class PaymentMethod extends Model
 {
-    use SoftDeletes;
     use HasFactory;
+    use SoftDeletes;
+    use PaymentMethodRequests;
 
     protected $guarded = ['id'];
 
     protected $hidden = [
         'token',
-        'exp_month',
-        'exp_year',
+        'details',
     ];
-
-    protected $appends = ['expiration_date'];
-
-    /**
-     * The payment method's pre-configured gateway.
-     *
-     * @var \rkujawa\LaravelPaymentGateway\PaymentGateway
-     */
-    private $paymentGateway;
 
     /**
      * Create a new factory instance for the model.
@@ -38,35 +31,6 @@ class PaymentMethod extends Model
     protected static function newFactory()
     {
         return PaymentMethodFactory::new();
-    }
-
-    public function setLastDigitsAttribute($value)
-    {
-        $this->attributes['last_digits'] = substr(preg_replace('/[^0-9]/', '', $value), -4);
-    }
-
-    public function setFirstNameAttribute($value)
-    {
-        if (!is_null($value)) {
-            $this->attributes['first_name'] = ucwords(strtolower(trim($value)));
-        }
-    }
-
-    public function setLastNameAttribute($value)
-    {
-        if (!is_null($value)) {
-            $this->attributes['last_name'] = ucwords(strtolower(trim($value)));
-        }
-    }
-
-    public function getCardholderAttribute()
-    {
-        return $this->first_name . ' ' . $this->last_name;
-    }
-
-    public function getExpirationDateAttribute()
-    {
-        return $this->getExpirationDate();
     }
 
     public function getProviderAttribute()
@@ -92,61 +56,5 @@ class PaymentMethod extends Model
     public function transactions()
     {
         return $this->hasMany(config('payment.models.' . PaymentTransaction::class, PaymentTransaction::class));
-    }
-
-    public function getExpirationDate(string $separator = '/', $yearFirst = false)
-    {
-        if ($yearFirst) {
-            return $this->exp_year . $separator . $this->exp_month;
-        }
-
-        return $this->exp_month . $separator . $this->exp_year;
-    }
-
-    /**
-     * Retrieve the payment method's configured gateway.
-     *
-     * @return \rkujawa\LaravelPaymentGateway\PaymentGateway
-     */
-    public function getGatewayAttribute()
-    {
-        if (! isset($this->paymentGateway)) {
-            $this->paymentGateway = (new PaymentGateway)
-                ->provider($this->wallet->provider)
-                ->merchant($this->wallet->merchant);
-        }
-        
-        return $this->paymentGateway;
-    }
-
-    /**
-     * Request the payment method details from the provider.
-     *
-     * @return \rkujawa\LaravelPaymentGateway\Contracts\PaymentManagerResponse
-     */
-    public function requestDetails()
-    {
-        return $this->gateway->getPaymentMethod($this);
-    }
-
-    /**
-     * Request the provider to update the payment method's details.
-     *
-     * @param array|mixed $data
-     * @return \rkujawa\LaravelPaymentGateway\Contracts\PaymentManagerResponse
-     */
-    public function requestUpdate($data)
-    {
-        return $this->gateway->updatePaymentMethod($this, $data);
-    }
-
-    /**
-     * Request the provider to remove the payment method from their system.
-     *
-     * @return \rkujawa\LaravelPaymentGateway\Contracts\PaymentManagerResponse
-     */
-    public function requestRemoval()
-    {
-        return $this->gateway->deletePaymentMethod($this);
     }
 }
